Move fake comm location
